backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pyotp
pikepdf
reportlab
python-dotenv
aiofiles
pytest
httpx
